Move `ValidatorGenerator` to the `generate` package
